+++ R/sfc.R
@@ -496,6 +496,9 @@ typed_empty = function(cls, ncol = 2, dim = "XY") {
 #'
 #' retrieve coordinates in matrix form
 #' @param x object of class sf, sfc or sfg
+#' @param round logical; if `TRUE` and input object has a `precision` attribute
+#'   different from 0, then its coordinates are rounded to the chosen precision
+#'   level before building the output matrix. See also `st_precision()`.
 #' @param ... ignored
 #' @return matrix with coordinates (X, Y, possibly Z and/or M) in rows, possibly followed by integer indicators \code{L1},...,\code{L3} that point out to which structure the coordinate belongs; for \code{POINT} this is absent (each coordinate is a feature), for \code{LINESTRING} \code{L1} refers to the feature, for \code{MULTILINESTRING} \code{L1} refers to the part and \code{L2} to the simple feature, for \code{POLYGON} \code{L1} refers to the main ring or holes and \code{L2} to the simple feature, for \code{MULTIPOLYGON} \code{L1} refers to the main ring or holes, \code{L2} to the ring id in the \code{MULTIPOLYGON}, and \code{L3} to the simple feature.
 #' 
@@ -506,18 +509,35 @@ typed_empty = function(cls, ncol = 2, dim = "XY") {
 #' feature and \code{L2} refers to the component \code{POLYGON}.
 #' 
 #' @export
-st_coordinates = function(x, ...) UseMethod("st_coordinates")
+#' @examples
+#' x = st_sfc(st_point(c(1.234, 1.234)), crs = 3003) # units: m
+#' st_coordinates(x)
+#' 
+#' st_precision(x) = 1e1
+#' st_coordinates(x, round = TRUE) # round 1 decimal place 
+#' 
+#' st_precision(x) = units::set_units(1, cm)
+#' st_coordinates(x, round = TRUE) # round to cm
+st_coordinates = function(x, round = FALSE, ...) UseMethod("st_coordinates")
 
 #' @export
-st_coordinates.sf = function(x, ...) st_coordinates(st_geometry(x))
+st_coordinates.sf = function(x, round = FALSE, ...) st_coordinates(st_geometry(x), round = round)
 
 #' @export
-st_coordinates.sfg = function(x, ...) st_coordinates(st_geometry(x))
+st_coordinates.sfg = function(x, round = FALSE, ...) st_coordinates(st_geometry(x), round = round)
 
 #' @export
-st_coordinates.sfc = function(x, ...) {
+st_coordinates.sfc = function(x, round = FALSE, ...) {
 	if (length(x) == 0)
 		return(matrix(nrow = 0, ncol = 2))
+	
+	# See the discussion in #2620 and #2622
+	if (round && st_precision(x) != 0) {
+		# NB: The following does not restore CRS, bbox, precision, ... of x but that
+		# shouldn't be relevant since we are only interested in the underlying
+		# representation of coordinates
+		x = st_as_sfc(st_as_binary(x, precision = st_precision(x)))
+	}
 
 	ret = switch(class(x)[1],
 		sfc_POINT = matrix(unlist(x, use.names = FALSE), nrow = length(x), byrow = TRUE,